+++ bubblegum/__init__.py
@@ -42,4 +42,4 @@ __all__ = [
     "__version__",
 ]
 
-__version__ = "0.0.6a72"
+__version__ = "0.0.6a73"